+++ v4/localenv.go
@@ -314,6 +314,15 @@ func (e *LocalEnv) GetPositionalParamCount() int {
 	return getPositionalParamCount(e)
 }
 
+// GetPositionalParams returns the (emulated) value of the UNIX
+// shell special parameter $@.
+//
+// It ignores any $@ that has been set in the environment, and builds
+// the list up using the value of $#.
+func (e *LocalEnv) GetPositionalParams() []string {
+	return getPositionalParams(e)
+}
+
 // ReplacePositionalParams sets $1, $2 etc etc to the given values.
 //
 // Any existing positional parameters are deleted.